updated file and folder resolver

+++ src/files/files.resolver.ts
@@ -6,6 +6,8 @@ import { GetManyFileInput } from './inputs/get-many.input';
 import { FilesService } from './files.service';
 import { JwtAuthGuard } from './../auth/guards/jwt-auth.guard';
 import { File } from './models/files.model';
+import { Auth } from './../auth/auth.decorator';
+import { User } from './../users/models/user.model';
 
 @Resolver()
 @UseGuards(JwtAuthGuard)
@@ -13,8 +15,8 @@ export class FilesResolver {
   constructor(private readonly service: FilesService) {}
 
   @Query(() => File)
-  async getOne(@Args('id') _id: string) {
-    const file = await this.service.getOne(_id);
+  async getOne(@Args('id') _id: string, @Auth() auth: User) {
+    const file = await this.service.getOne({ _id, user_id: auth._id });
 
     if (!file) {
       return new NotFoundException();
@@ -24,30 +26,36 @@ export class FilesResolver {
   }
 
   @Query(() => [File])
-  async getMany(@Args('params') params: GetManyFileInput) {
-    return await this.service.getMany(params);
+  async getMany(@Args('params') params: GetManyFileInput, @Auth() auth: User) {
+    return await this.service.getMany({ ...params, user_id: auth._id });
   }
 
   @Mutation(() => File)
-  async createOne(@Args('createFileInput') body: CreateFileInput) {
-    return await this.service.createOne(body);
+  async createOne(
+    @Args('createFileInput') body: CreateFileInput,
+    @Auth() auth: User,
+  ) {
+    return await this.service.createOne({ ...body, user_id: auth._id });
   }
 
   @Mutation(() => File)
-  async updateOne(@Args('updateFileInput') body: UpdateFileInput) {
+  async updateOne(
+    @Args('updateFileInput') body: UpdateFileInput,
+    @Auth() auth: User,
+  ) {
     const { _id } = body;
 
-    const file = await this.service.getOne(_id);
+    const file = await this.service.getOne({ _id, user_id: auth._id });
 
     if (!file) {
       return new NotFoundException();
     }
 
-    return await this.service.updateOne(body);
+    return await this.service.updateOne({ ...body, user_id: auth._id });
   }
 
   @Mutation(() => File)
-  async deleteOne(@Args('id') _id: string) {
-    return await this.service.deleteOne(_id);
+  async deleteOne(@Args('id') _id: string, @Auth() auth: User) {
+    return await this.service.deleteOne({ _id, user_id: auth._id });
   }
 }
